add dummy data for search endpoint test
