+++ tests/blueprint_specific_tests/search_bp_tests/test_search_endpoints.py
@@ -17,22 +17,32 @@ def test_pipelines_route(test_client):
     assert res.status_code == 200
 
 
-def test_pipelines_route(test_client):
+def test_dataset_search_route(session, new_dataset, new_dataset_stats, test_client):
     """
     GIVEN calling the route "/dataset-search"
     WHEN no user is logged in
     THEN should return success code
+
+    TODO: mock the datasets table to provide a valid return object
     """
-    #res = test_client.get("/dataset-search")
-    #assert res.status_code == 200
 
-    assert True
+    session.add(new_dataset)
+    session.commit()
 
-def test_pipelines_route(test_client):
+    session.add(new_dataset_stats)
+    session.commit()
+
+    res = test_client.get("/dataset-search")
+    assert res.status_code == 200
+
+
+def test_dataset_route(test_client):
     """
     GIVEN calling the route "/dataset"
     WHEN no user is logged in
     THEN should return success code
+
+    TODO: mock the datasets table to provide a valid return object
     """
     #res = test_client.get("/dataset")
     #assert res.status_code == 200
